requirements.txt
flask
flask_restplus 
deepspeech-keras